fix: 모델 분리
