+++ service/user.py
@@ -9,12 +9,12 @@ from service.auth import (
     hash_password,
 )
 from fastapi import Depends
-from models import models
+from models import users
 from sqlalchemy.orm import Session
 from database import SessionLocal
 from fastapi.security import OAuth2PasswordRequestForm
 from datetime import timedelta
-from models.models import CreateUser
+from models.users import CreateUser
 
 
 def get_db():
@@ -26,7 +26,7 @@ def get_db():
 
 
 async def create_user(user: CreateUser, db: Session = Depends(get_db)):
-    new_user = models.User()
+    new_user = user.User()
     new_user.username = user.username
     new_user.email = user.email
     new_user.password = hash_password(user.password)
